pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sleep-esi"
version = "0.1.0"
description = "NYC Environmental Sleep Index (CD59) - Geospatial analysis pipeline"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Sleep ESI Team"}
]

dependencies = [
    "geopandas>=0.14",
    "pandas>=2.1",
    "pyarrow>=14.0",
    "numpy>=1.26",
    "pyyaml>=6.0",
    "rasterio>=1.3",
    "rasterstats>=0.19",
    "scipy>=1.11",
    "statsmodels>=0.14",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov",
    "black",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "smoke: marks tests as smoke tests (run with pytest -m smoke)",
]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]

[tool.black]
line-length = 100

